==> notebooks/01_ingestion.ipynb <==
import pandas as pd
from pathlib import Path

DATA = Path("../data/bronze")

train_path = DATA / "train.csv"
store_path = DATA / "store.csv"

train = pd.read_csv(train_path)
store = pd.read_csv(store_path)

print("train:", train.shape)
print("store:", store.shape)
train.head()

train.info()
train.isna().mean().sort_values(ascending=False).head(10)

df = train.merge(store, on="Store", how="left")
print("merged:", df.shape)
df.head()
